requirements.txt
transformers
pandas
numpy
torch
scikit-learn
re
ast
XLMRobertaTokenizer
matplotlib.pyplot
AutoTokenizer
datasets
Dataset
AutoModelForSequenceClassification
Trainer
TrainingArguments
sklearn
model_selection
train_test_split
sklearn.metrics
f1_score
precision_score
recall_score